+++ eslint.config.js
@@ -1,0 +1,3 @@
+import baseConfig from './rules/esnext.js';
+
+export default baseConfig;