+++ rules/optional/index.js
@@ -1,11 +1,23 @@
-const merge = require('deepmerge');
-const { isDependency } = require('@aduth/is-dependency');
-const jsdoc = require('./jsdoc.js');
-const prettier = require('./prettier.js');
-const typescript = require('./typescript.js');
+import { isDependency } from '@aduth/is-dependency';
 
-module.exports = [jsdoc, prettier, typescript].reduce(
-	(result, { dependencies, config }) =>
-		isDependency(dependencies) ? merge(result, config) : result,
-	{}
-);
+const isLocalDependency = (name) =>
+	isDependency(name, { fields: ['dependencies', 'devDependencies'] });
+
+const optionals = [];
+
+if (isLocalDependency('eslint-plugin-jsdoc')) {
+	const { default: jsdoc } = await import('./jsdoc.js');
+	optionals.push(...jsdoc);
+}
+
+if (isLocalDependency(['eslint-plugin-prettier', 'eslint-config-prettier'])) {
+	const { default: prettier } = await import('./prettier.js');
+	optionals.push(...prettier);
+}
+
+if (isLocalDependency('typescript-eslint')) {
+	const { default: typescript } = await import('./typescript.js');
+	optionals.push(...typescript);
+}
+
+export default optionals;